strong wip, getting somewhere...
