+++ compiler/wasm/.eslintignore
@@ -1,1 +1,3 @@
 node_modules
+noir_wasm/pkg
+test/
