--- compiler/wasm/index.js
@@ -1,3 +1,0 @@
-const module = import('./noir_wasm/pkg');
-
-const { compile } = module;
