+++ compiler/wasm/index.ts
@@ -1,0 +1,7 @@
+// import { compile } from './noir_wasm/pkg';
+
+const compile = async () => {
+  console.log('hey');
+};
+
+export { compile };
