+++ compiler/wasm/noir_wasm/src/compile.rs
@@ -279,29 +279,30 @@ fn preprocess_contract(contract: CompiledContract) -> CompileResult {
     CompileResult::Contract { contract: preprocessed_contract, debug: debug_artifact }
 }
 
-cfg_if::cfg_if! {
-    if #[cfg(target_os = "wasi")] {
-        fn get_non_stdlib_asset(path_to_file: &Path) -> std::io::Result<String> {
-            std::fs::read_to_string(path_to_file)
-        }
-    } else {
-        use std::io::{Error, ErrorKind};
-
-        #[wasm_bindgen(module = "@noir-lang/source-resolver")]
-        extern "C" {
-            #[wasm_bindgen(catch)]
-            fn read_file(path: &str) -> Result<String, JsValue>;
-        }q
-
-        fn get_non_stdlib_asset(path_to_file: &Path) -> std::io::Result<String> {
-            let path_str = path_to_file.to_str().unwrap();
-            match read_file(path_str) {
-                Ok(buffer) => Ok(buffer),
-                Err(_) => Err(Error::new(ErrorKind::Other, "could not read file using wasm")),
-            }
-        }
-    }
+fn get_non_stdlib_asset(path_to_file: &Path) -> std::io::Result<String> {
+    std::fs::read_to_string(path_to_file)
 }
+
+// cfg_if::cfg_if! {
+// if #[cfg(target_os = "wasi")] {
+// } else {
+//     use std::io::{Error, ErrorKind};
+
+// #[wasm_bindgen(module = "@noir-lang/source-resolver")]
+// extern "C" {
+//     #[wasm_bindgen(catch)]
+//     fn read_file(path: &str) -> Result<String, JsValue>;
+// }q
+
+// fn get_non_stdlib_asset(path_to_file: &Path) -> std::io::Result<String> {
+//     let path_str = path_to_file.to_str().unwrap();
+//     match read_file(path_str) {
+//         Ok(buffer) => Ok(buffer),
+//         Err(_) => Err(Error::new(ErrorKind::Other, "could not read file using wasm")),
+//     }
+// }
+// }
+// }
 
 #[cfg(test)]
 mod test {
