+++ compiler/wasm/test/shared.cjs
@@ -1,0 +1,8 @@
+module.exports = {
+  simpleScriptSourcePath: '../../fixtures/simple/noir-script/src/main.nr',
+  simpleScriptExpectedArtifact: '../../fixtures/simple/noir-script/target/noir_wasm_testing.json',
+  depsScriptSourcePath: '../../fixtures/deps/noir-script/src/main.nr',
+  depsScriptExpectedArtifact: '../../fixtures/deps/noir-script/target/noir_wasm_testing.json',
+  libASourcePath: '../../fixtures/deps/lib-a/src/lib.nr',
+  libBSourcePath: '../../fixtures/deps/lib-b/src/lib.nr',
+};
