--- compiler/wasm/tsconfig.cjs.json
@@ -1,18 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "outDir": "dist/types",
-    "module": "CommonJS",
-    "strict": true,
-    "experimentalDecorators": true,
-    "esModuleInterop": true,
-    "noImplicitAny": true,
-    "removeComments": false,
-    "preserveConstEnums": true,
-    "sourceMap": true,
-    "resolveJsonModule": true,
-    "importHelpers": true,
-    "allowSyntheticDefaultImports": true
-  },
-  "include": ["src/**/*"]
-}
