--- compiler/wasm/tsconfig.esm.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "node",
-    "outDir": "dist/types",
-    "module": "ESNext",
-    "strict": true,
-    "experimentalDecorators": true,
-    "esModuleInterop": true,
-    "noImplicitAny": true,
-    "removeComments": false,
-    "preserveConstEnums": true,
-    "sourceMap": true,
-    "resolveJsonModule": true,
-    "importHelpers": true,
-  },
-  "include": ["src/**/*"]
-}
