+++ compiler/wasm/tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "compilerOptions": {
+    "moduleResolution": "node",
+    "outDir": "dist/types",
+    "strict": true,
+    "experimentalDecorators": true,
+    "noImplicitAny": true,
+    "removeComments": false,
+    "preserveConstEnums": true,
+    "sourceMap": true,
+    "resolveJsonModule": true,
+    "importHelpers": true,
+    "allowSyntheticDefaultImports": true,
+    "esModuleInterop": true
+  },
+  "exclude": ["test"],
+  "include": ["noir_wasm/**/*.ts"],
+}
